+++ templates/hexarch/Infrastructure/Infrastructure.Adapters/Db/Migrations/20260114142258_AddDeadline.cs
@@ -32,12 +32,24 @@ namespace Arbeidstilsynet.HexagonalArchitectureTemplateDocker.Infrastructure.Ada
                 type: "timestamp with time zone",
                 nullable: false
             );
+            
+            // Fix typo in column name
+            migrationBuilder.RenameColumn(
+                name: "Organisajonsnummer",
+                table: "Saker",
+                newName: "Organisasjonsnummer");
+            
         }
 
         /// <inheritdoc />
         protected override void Down(MigrationBuilder migrationBuilder)
         {
             migrationBuilder.DropColumn(name: "Deadline", table: "Saker");
+            
+            migrationBuilder.RenameColumn(
+                name: "Organisasjonsnummer",
+                table: "Saker",
+                newName: "Organisajonsnummer");
         }
     }
 }